pyproject.toml
[project]
name = "cloudeagle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "browser-use>=0.4.2",
    "deepseek>=1.0.0",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.17",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.3",
    "openai>=1.93.0",
    "playwright>=1.53.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "webdriver-manager>=4.0.2",
]
